requirements.txt
SpeechRecognition
PySerial
Colorama
